# apps/backend/cmd/docs/swagger.yaml
basePath: /api
definitions:
...
  admingo_internal_modules_rbac_dto.PermissionDTO:
    properties:
      code:
        type: string
      path:
        type: string
    type: object
  admingo_internal_modules_rbac_dto.RoleDTO:
    properties:
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/admingo_internal_modules_rbac_dto.PermissionDTO'
        type: array
    required:
    - name
    type: object
...
  admingo_internal_pkg_response.Response-array_admingo_internal_modules_rbac_dto_RoleDTO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/admingo_internal_modules_rbac_dto.RoleDTO'
        type: array
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend service for AdminGO.
  title: AdminGO API
  version: "1.0"
paths:
...
  /api/sys/menu/roles:
    get:
      consumes:
      - application/json
      description: 分页获取角色及其权限
      parameters:
      - description: 页码 (默认 1)
        in: query
        name: page
        type: integer
      - description: 每页数量 (默认 10)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回角色列表
          schema:
            $ref: '#/definitions/admingo_internal_pkg_response.Response-array_admingo_internal_modules_rbac_dto_RoleDTO'
      summary: 获取角色列表（包含权限）
      tags:
      - 角色管理
swagger: "2.0"